only run splotch if file DNE

+++ workflows/run_csplotch.wdl
@@ -94,15 +94,18 @@ task run_splotch {
         do
             GENE_IDX=${ALL_GENES[$IDX]}
             GENE_DIR=$((GENE_IDX / 100))
-            GENE_FILE=~{csplotch_input_dir_stripped}/$GENE_DIR/data_$GENE_IDX.R
-            mkdir -p ./data_directory/$GENE_DIR
-            gsutil cp $GENE_FILE ./data_directory/$GENE_DIR/
-            
-            mkdir -p ./csplotch_outputs/$GENE_DIR/
-            
-            splotch -g $GENE_IDX -d ./data_directory -o ./csplotch_outputs -b $SPLOTCH_BIN -n ~{num_samples} -c ~{num_chains} -s
-            
-            gsutil cp ./csplotch_outputs/$GENE_DIR/combined_$GENE_IDX.hdf5 ~{csplotch_output_dir_stripped}/$GENE_DIR/combined_$GENE_IDX.hdf5
+            SUMMARY_EXISTS=`gsutil -q stat gs://fc-c230706b-9a02-457b-b086-4371eef75869/nmj_assembloid/csplotch_outputs/$GENE_DIR/combined_$GENE_IDX.hdf5; echo $?`
+            if [ $SUMMARY_EXISTS -ne 0 ]; then
+                GENE_FILE=~{csplotch_input_dir_stripped}/$GENE_DIR/data_$GENE_IDX.R
+                mkdir -p ./data_directory/$GENE_DIR
+                gsutil cp $GENE_FILE ./data_directory/$GENE_DIR/
+                
+                mkdir -p ./csplotch_outputs/$GENE_DIR/
+                
+                splotch -g $GENE_IDX -d ./data_directory -o ./csplotch_outputs -b $SPLOTCH_BIN -n ~{num_samples} -c ~{num_chains} -s
+                
+                gsutil cp ./csplotch_outputs/$GENE_DIR/combined_$GENE_IDX.hdf5 ~{csplotch_output_dir_stripped}/$GENE_DIR/combined_$GENE_IDX.hdf5
+            fi
             
         done
